chore: moved losses to losses.py for re-use
